oh boy, all sorts of savagery
