+++ src/Language/Fixpoint/Types.hs
@@ -189,6 +189,7 @@ import           Text.Printf               (printf)
 import           Language.Fixpoint.Config
 import           Language.Fixpoint.Misc
 import           Language.Fixpoint.Names
+import           Language.Fixpoint.Errors
 import           Text.Parsec.Pos
 import           Text.PrettyPrint.HughesPJ
 
@@ -196,15 +197,6 @@ import           Data.Array                hiding (indices)
 import qualified Data.HashMap.Strict       as M
 import qualified Data.HashSet              as S
 
-
-class Fixpoint a where
-  toFix    :: a -> Doc
-
-  simplify :: a -> a
-  simplify =  id
-
-showFix :: (Fixpoint a) => a -> String
-showFix =  render . toFix
 
 traceFix     ::  (Fixpoint a) => String -> a -> a
 traceFix s x = trace ("\nTrace: [" ++ s ++ "] : " ++ showFix x) x
@@ -451,6 +443,9 @@ data Expr = ESym !SymConst
 elit :: Located Symbol -> Sort -> Expr
 elit l s = ECon $ L (symbolText $ val l) s
 
+instance Fixpoint Int where
+  toFix = tshow
+
 instance Fixpoint Integer where
   toFix = integer
 
@@ -503,10 +498,146 @@ instance Fixpoint Expr where
   toFix (ECst e so)    = parens $ toFix e   <+> text " : " <+> toFix so
   toFix (EBot)         = text "_|_"
 
-----------------------------------------------------------
---------------------- Predicates -------------------------
-----------------------------------------------------------
 
+instance PPrint Constant where
+  pprint = toFix
+
+instance PPrint Brel where
+  pprint Eq = text "=="
+  pprint Ne = text "/="
+  pprint r  = toFix r
+
+instance PPrint Bop where
+  pprint  = toFix
+
+instance PPrint Sort where
+  pprint = toFix
+
+instance PPrint KVar where
+  pprint (KV x) = text "$" <> pprint x
+
+instance PPrint SymConst where
+  pprint (SL x) = doubleQuotes $ text $ T.unpack x
+
+-- | Wrap the enclosed 'Doc' in parentheses only if the condition holds.
+parensIf True  = parens
+parensIf False = id
+
+-- NOTE: The following Expr and Pred printers use pprintPrec to print
+-- expressions with minimal parenthesization. The precedence rules are somewhat
+-- fragile, and it would be nice to have them directly tied to the parser, but
+-- the general idea is (from lowest to highest precedence):
+--
+-- 1 - if-then-else
+-- 2 - => and <=>
+-- 3 - && and ||
+-- 4 - ==, !=, <, <=, >, >=
+-- 5 - mod
+-- 6 - + and -
+-- 7 - * and /
+-- 8 - function application
+--
+-- Each printer `p` checks whether the precedence of the context is greater than
+-- its own precedence. If so, the printer wraps itself in parentheses. Then it
+-- sets the contextual precedence for recursive printer invocations to
+-- (prec p + 1).
+
+opPrec Mod   = 5
+opPrec Plus  = 6
+opPrec Minus = 6
+opPrec Times = 7
+opPrec Div   = 7
+
+instance PPrint Expr where
+  pprintPrec _ (ESym c)        = pprint c
+  pprintPrec _ (ECon c)        = pprint c
+  pprintPrec _ (EVar s)        = pprint s
+  pprintPrec _ (EBot)          = text "_|_"
+  pprintPrec z (ENeg e)        = parensIf (z > zn) $
+                                   text "-" <> pprintPrec (zn+1) e
+    where zn = 2
+  pprintPrec z (EApp f es)     = parensIf (z > za) $
+                                   intersperse empty $
+                                     pprint f : (pprintPrec (za+1) <$> es)
+    where za = 8
+  pprintPrec z (EBin o e1 e2)  = parensIf (z > zo) $
+                                   pprintPrec (zo+1) e1 <+>
+                                   pprint o             <+>
+                                   pprintPrec (zo+1) e2
+    where zo = opPrec o
+  pprintPrec z (EIte p e1 e2)  = parensIf (z > zi) $
+                                   text "if"   <+> pprintPrec (zi+1) p  <+>
+                                   text "then" <+> pprintPrec (zi+1) e1 <+>
+                                   text "else" <+> pprintPrec (zi+1) e2
+    where zi = 1
+  pprintPrec _ (ECst e so)     = parens $ pprint e <+> text ":" <+> pprint so
+
+instance PPrint Pred where
+  pprintPrec _ PTop            = text "???"
+  pprintPrec _ PTrue           = trueD
+  pprintPrec _ PFalse          = falseD
+  pprintPrec z (PBexp e)       = pprintPrec z e
+  pprintPrec z (PNot p)        = parensIf (z > zn) $
+                                   text "not" <+> pprintPrec (zn+1) p
+    where zn = 8
+  pprintPrec z (PImp p1 p2)    = parensIf (z > zi) $
+                                   (pprintPrec (zi+1) p1) <+>
+                                   text "=>"              <+>
+                                   (pprintPrec (zi+1) p2)
+    where zi = 2
+  pprintPrec z (PIff p1 p2)    = parensIf (z > zi) $
+                                   (pprintPrec (zi+1) p1) <+>
+                                   text "<=>"             <+>
+                                   (pprintPrec (zi+1) p2)
+    where zi = 2
+  pprintPrec z (PAnd ps)       = parensIf (z > za) $
+                                   pprintBin (za+1) trueD  andD ps
+    where za = 3
+  pprintPrec z (POr  ps)       = parensIf (z > zo) $
+                                   pprintBin (zo+1) falseD orD  ps
+    where zo = 3
+  pprintPrec z (PAtom r e1 e2) = parensIf (z > za) $
+                                   pprintPrec (za+1) e1 <+>
+                                   pprint r             <+>
+                                   pprintPrec (za+1) e2
+    where za = 4
+  pprintPrec _ (PAll xts p)    = text "forall" <+> toFix xts <+> text "." <+> pprint p
+  pprintPrec _ (PExist xts p)  = text "exists" <+> toFix xts <+> text "." <+> pprint p
+  pprintPrec _ p@(PKVar {})    = toFix p
+
+trueD  = text "true"
+falseD = text "false"
+andD   = text " &&"
+orD    = text " ||"
+
+pprintBin _ b _ [] = b
+pprintBin z _ o xs = intersperse o $ pprintPrec z <$> xs
+
+instance PPrint Reft where
+  pprint r@(Reft (_,ra))
+    | isTauto r        = text "true"
+    | otherwise        = {- intersperse comma -} pprintBin z trueD andD flat
+    where
+      flat = flattenRefas [ra]
+      z    = if length flat > 1 then 3 else 0
+
+instance PPrint SortedReft where
+  pprint (RR so (Reft (v, ras)))
+    = braces
+    $ pprint v <+> text ":" <+> toFix so <+> text "|" <+> pprint ras
+
+instance PPrint a => PPrint (Located a) where
+  pprint (Loc _ _ x) = pprint x
+
+
+instance PTable (SInfo a) where
+  ptable fi = DocTable [ (text "# Sub Constraints", pprint $ length $ cm fi)
+                       , (text "# WF  Constraints", pprint $ length $ ws fi)
+                       ]
+
+--------------------------------------------------------------------------------
+-- | Predicates ----------------------------------------------------------------
+--------------------------------------------------------------------------------
 
 data Pred = PTrue
           | PFalse
@@ -969,12 +1100,6 @@ instance Monoid (Result a) where
 -- instance Functor Result where
 --   fmap f (Result x y) = Result (fmap (fmap f) x) y
 
-data FixResult a = Crash [a] String
-                 | Safe
-                 | Unsafe ![a]
-                 | UnknownError !String
-                   deriving (Data, Typeable, Show, Generic)
-
 type FixSolution = M.HashMap KVar Pred
 
 instance Eq a => Eq (FixResult a) where
@@ -1061,9 +1186,6 @@ toFixMeta k v = text "// META" <+> k <+> text ":" <+> v
 
 pprId (Just i)  = text "id" <+> tshow i
 pprId _         = text ""
-
-instance Fixpoint Int where
-  toFix = tshow
 
 -------------------------------------------------------
 ------------------- Substitutions ---------------------
@@ -1667,12 +1789,6 @@ editDistance xs ys = table ! (m,n)
 -----------------------------------------------------------------------------
 -- | Located Values ---------------------------------------------------------
 -----------------------------------------------------------------------------
-
-data Located a = Loc { loc  :: !SourcePos -- ^ Start Position
-                     , locE :: !SourcePos -- ^ End Position
-                     , val  :: a
-                     } deriving (Data, Typeable, Generic)
-
 instance (IsString a) => IsString (Located a) where
   fromString = dummyLoc . fromString
