+++ src/Language/Fixpoint/Types/Visitor.hs
@@ -1,9 +1,10 @@
-{-# LANGUAGE TupleSections #-}
+{
+# LANGUAGE TupleSections #-}
 {-# LANGUAGE PatternGuards #-}
 {-# LANGUAGE TypeSynonymInstances #-}
 {-# LANGUAGE FlexibleInstances #-}
 
-module Language.Fixpoint.Visitor (
+module Language.Fixpoint.Types.Visitor (
   -- * Visitor
      Visitor (..)
