+++ src/Language/Fixpoint/Utils/Statistics.hs
@@ -2,7 +2,7 @@
 
 -- | This module implements functions that print out
 --   statistics about the constraints.
-module Language.Fixpoint.Statistics (statistics) where
+module Language.Fixpoint.Utils.Statistics (statistics) where
 
 import           Control.DeepSeq
 import           GHC.Generics
